[Encoder] fix a small range of slice idx

+++ codec/encoder/core/inc/slice.h
@@ -168,7 +168,7 @@ SMVUnitXY	sMvc[5];
 uint8_t		uiMvcNum;
 uint8_t		sScaleShift;
 
-uint8_t		uiSliceIdx;
+uint32_t		uiSliceIdx;
 bool		bSliceHeaderExtFlag; // Indicate which slice header is used, avc or ext?
 uint8_t		uiLastMbQp;		// stored qp for last mb coded, maybe more efficient for mb skip detection etc.
